Begun work on a new PonyEditor, with a tabbed interface.
Fixed screensaver mode to be forgiving of unset color/image when these modes are used. The background form also uses mouse movement to trigger close.
Added factory constructor for AnimatedImage of type BitmapFrame.
Fixed a change introduced in GifImage that resulted in a non-terminating loop.
Made some fixes to IdleWorker disposal.
Made some changes to improve reloading after returning from the editor.

+++ C# Desktop Ponies/SpriteManagement/GifImage.cs
@@ -1663,10 +1663,10 @@
             int stackIndex = 0;
 
             // Initialize table with root characters.
-            for (byte i = 0; i < clearCode; i++)
+            for (int i = 0; i < clearCode; i++)
             {
                 prefix[i] = 0;
-                suffix[i] = i;
+                suffix[i] = (byte)i;
             }
             #endregion
